pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vera-benchmark"
version = "1.0.0"
description = "VERA: Voice-Enabled Reasoning Assessment Benchmark for Evaluating Reasoning Resilience in Voice Agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "VERA Team", email = "vera@example.com"}
]
keywords = ["voice", "reasoning", "benchmark", "ai", "evaluation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "numpy>=1.26.0",
    "scipy>=1.7.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "datasets>=2.0.0",
    "librosa>=0.9.0",
    "soundfile>=0.12.0",
    "webrtcvad>=2.0.10",
    "websockets>=12.0",
    "azure-cognitiveservices-speech>=1.30.0",
    "requests>=2.25.0",
    "httpx>=0.24.0",
    "tqdm>=4.60.0",
    "jsonlines>=3.0.0",
    "tiktoken>=0.4.0",
    "python-dotenv>=0.19.0",
    "rich>=12.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "pandas>=1.3.0",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pydub>=0.25.1",
    "nemo-toolkit[asr]>=1.23.0",
    "google-genai>=1.36.0",
    "websocket-client>=1.8.0",
    "vllm>=0.1.2",
    "groq>=0.31.1",
    "sphn>=0.2.0",
    "python-socketio[client]>=5.13.0",
    "dotenv>=0.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
    "pre-commit>=2.17.0",
]

[project.urls]
Homepage = "https://github.com/anonymous/vera"
Documentation = "https://github.com/anonymous/vera"
Repository = "https://github.com/anonymous/vera.git"
"Bug Tracker" = "https://github.com/anonymous/vera/issues"

[project.scripts]
vera = "vera.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "*.tests*", "models*"]

[tool.setuptools.package-data]
vera = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=vera --cov-report=html --cov-report=term-missing"
